move loaderExtras file location
